refactor: move files to helpers folder
